tests/wpt/web-platform-tests/app-history/app-history-entry/entries-when-inactive.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  const first_entry = appHistory.entries()[0];
  history.pushState(1, "", "#1");
  assert_equals(appHistory.entries()[0], first_entry);
  history.back();
  window.onpopstate = t.step_func_done(() => {
    const second_entry = appHistory.entries()[1];
    history.replaceState(0, "", "#0");
    assert_equals(appHistory.entries()[1], second_entry);
  });
}, "A non-active entry in appHistory.entries() should not be modified when a different entry is modified");
</script>
